+++ apps/dashboard/src/lib/constants.ts
@@ -1,5 +1,5 @@
 // These should match the rust constants in fig_util::consts
-export const PRODUCT_NAME = "Q";
+export const PRODUCT_NAME = "Amazon Q";
 export const CLI_BINARY_NAME = "q";
 
 export const USER_MANUAL_URL =