+++ lib/fig_integrations/src/intellij/macos.rs
@@ -15,7 +15,7 @@ use crate::{
 
 const PLUGIN_PREFIX: &str = "q-for-command-line-companion";
 
-/// Version of the Q Intellij plugin. This should match the version in
+/// Version of the Amazon Q Intellij plugin. This should match the version in
 /// [build.gradle](../../../../extensions/jetbrains/build.gradle).
 const PLUGIN_VERSION: &str = "1.0.2";
 
@@ -24,8 +24,8 @@ const OLD_PLUGIN_SLUGS: [&str; 2] = ["codewhisperer-for-command-line-companion",
 
 static PLUGIN_CONTENTS: &[u8] = include_bytes!("plugin.jar");
 
-/// Returns the identifier of the plugin - i.e., the directory name of the Q plugin stored under the
-/// Intellij `plugins/` directory.
+/// Returns the identifier of the plugin - i.e., the directory name of the Amazon Q plugin stored
+/// under the Intellij `plugins/` directory.
 fn plugin_slug() -> String {
     format!("{PLUGIN_PREFIX}-{PLUGIN_VERSION}")
 }